backend/pyproject.toml
[project]
name = "insightflow-backend"
version = "0.1.0"
description = "InsightFlow backend service"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "alembic>=1.13.2",
  "email-validator>=2.2.0",
  "fastapi>=0.115.0",
  "passlib[bcrypt]>=1.7.4",
  "psycopg2-binary>=2.9.9",
  "pydantic-settings>=2.4.0",
  "python-jose[cryptography]>=3.3.0",
  "python-multipart>=0.0.9",
  "sqlalchemy>=2.0.34",
  "uvicorn[standard]>=0.30.6",
]

[project.optional-dependencies]
dev = [
  "httpx>=0.27.2",
  "pytest>=8.3.2",
]

[tool.pytest.ini_options]
testpaths = ["tests"]

